MAGEP-42: Return non-zero exit code for CLI commands
Currently when "bin/magento conifgurator:run" is executed and fails, then an error message is printed to console output, but the command still exits with 0 code which is an indication of a successful command termination.
A new CommandFailedException indicates that further processing is not possible, and the an error has to be reported to the user.
This commit makes it return 1 for all ComponentException errors.